Modified stop condition to be on full passing of reward window and added more finely binned reward calculation

+++ include/dynamicTau/environment/bayes.hpp
@@ -40,7 +40,7 @@ class BayesFactors{
     std::vector<double> coeffSumB;
     
     // Reward parameters
-    double rewardStartIndex, rewardEndIndex;
+    int rewardStartIndex, rewardEndIndex;
     double targetCoupling;
     // Piecewise reward
     double belowThresholdValue, aboveThresholdValue, stepPunish;
